services/vision/requirements.txt
litestar==2.13.0
uvicorn[standard]==0.32.1
pydantic==2.10.3
python-dotenv==1.0.1
numpy==2.2.0
httpx==0.28.1